requirements.txt
boto3==1.9.106
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-constrainedfilefield==3.2.0
idna==2.7
moviepy==1.0.0
numpy==1.15.4
psycopg2-binary==2.7.6.1
python-magic-bin==0.4.14
pytz==2018.9
sagemaker==1.18.2
scikit-learn==0.20.2
scipy==1.2.0
torch==0.4.0
urllib3==1.24.1
wincertstore==0.2
